when Lazy is used, set flushEnabled=true
Issue #28 make sure to enable the global flushing after calling clear() when Lazy is used. In this case, the EntityManager is the most likely to be in a transaction context. This change works with https://github.com/42BV/beanmapper/issues/90, where the default for the BeanCollection is set to true. This means flushing will be called by default for BeanCollections operating in a Lazy "context".