#32 MergePair.beforeMerge must be untouched, not BeanMapped
When a MergePair is used, the beforeMerge is now retrieved and
immediately detached before the target is fetched. Take note that
this means that the original's proxies can no longer be resolved.
This restriction does not apply to the afterMerge instance, which
is still tied to its session as usual. The previous mechanism of
mapping the original (beforeMerge) to a variable using BeanMapper
has been scuttled.

EntityFinder has become a bit-more dev-friendly by taking note of
the classes it can return.

The constructor of MergedFormMethodArgumentResolver has changed so
that it can safely pass on EntityManager, even if null.

BeanConverters have been updated to the usage of BeanPropertyMatch
instead of BeanFieldMatch.